// types/test.ts
export enum ContentSources {
  AI = 'ai',
  Original = 'original',
  Programmatic = 'programmatic'
}

export enum ContentLengths {
  Longer = 'longer',
  Shorter = 'shorter',
}

export enum QuestionnaireTypes {
  Pre = 'pre',
  Post = 'post',
}

export enum TestSlugs {
    PRACTICE = 'practice',
    PUSH_NOTIFICATIONS = 'push-notifications',
    SEARCH_ENGINE = 'search-engine',
    EMAIL_INBOX = 'email-inbox',
    PRODUCT_LISTING = 'product-listing',
    MEETING_TRANSCRIPTION = 'meeting-transcription',
    PRESENTATION_SLIDE = 'presentation-slide',
}